gwpy.types: renamed core gwpy.data module
