+++ gwpy/frequencyseries/hist.py
@@ -27,7 +27,7 @@ if sys.version_info[0] < 3:
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
 
-from ..data import (Quantity, Array, Array2D)
+from ..types import (Quantity, Array, Array2D)
 from .core import FrequencySeries
 from ..detector import Channel
