+++ gwpy/plotter/core.py
@@ -694,11 +694,11 @@ class Plot(figure.Figure):
     @auto_refresh
     def add_array(self, array, projection, ax=None, newax=False,
                   sharex=None, sharey=None, **kwargs):
-        """Add a :class:`~gwpy.data.array.Array` to this plot
+        """Add a :class:`~gwpy.types.Array` to this plot
 
         Parameters
         ----------
-        array : :class:`~gwpy.data.array.Array`
+        array : :class:`~gwpy.types.Array`
             the `Array` to display
         projection : `str`
         ax : :class:`~gwpy.plotter.Axes`
