+++ gwpy/types/__init__.py
