+++ gwpy/types/array.py
