+++ gwpy/types/array2d.py
