+++ gwpy/types/io/__init__.py
