+++ gwpy/types/io/hdf5.py
