+++ gwpy/types/series.py
